requirements.txt
asgiref==3.10.0
cachetools==6.2.4
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.3.1
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
